--- a/tools/mtouch/Assembly.cs
+++ b/tools/mtouch/Assembly.cs
@@ -338,6 +338,17 @@ namespace Xamarin.Bundler {
 					compiler_flags.ReferenceSymbols (Target.GetRequiredSymbols (this, true));
 			}
 
+			if (App.EnableLLVMOnlyBitCode) {
+				// The AOT compiler doesn't optimize the bitcode so clang will do it
+				compiler_flags.AddOtherFlag ("-fexceptions");
+				var optimizations = App.GetLLVMOptimizations (this);
+				if (optimizations == null) {
+					compiler_flags.AddOtherFlag ("-O2");
+				} else if (optimizations.Length > 0) {
+					compiler_flags.AddOtherFlag (optimizations);
+				}
+			}
+
 			link_task = new LinkTask ()
 			{
 				Target = Target,