@rolfbjarne 2017-01-20
The test failure is not related, and will be fixed with PR #1541.